+++ src/langgraph/langgraph_agent.py
@@ -1,5 +1,3 @@
-
-
 import os
 from dotenv import load_dotenv
 from langchain_openai import ChatOpenAI
@@ -23,7 +21,7 @@ class LangGraphAgent:
         agent_dict = json.load(open(agent_path, "r"))
 
         if "langchain_config" not in agent_dict or not agent_dict["langchain_config"].get("use_langchain", False):
-            raise ValueError("Langchain is not configured properly in the agent config")
+            raise ValueError("You must have a langchain_config configuration in your agent file to use LangGraphAgent")
 
         if "config" not in agent_dict:
             raise KeyError("Missing required fields: config")
